move energy exchange methods to RadiosityFast.py
