+++ sparrowpy/classes/RadiosityFast.py
@@ -4,7 +4,6 @@ import pyfar as pf
 from sparrowpy.form_factor import form_factor
 from sparrowpy.radiosity_fast import (
     source_energy, receiver_energy, geometry)
-from sparrowpy.radiosity_fast import energy_exchange_order as ee_order
 try:
     import numba
     prange = numba.prange
@@ -182,7 +181,7 @@ class DirectionalRadiosityFast():
         energy_0, distance_0 = source_energy._init_energy_universal(
             source_position, patches_center, self.patches_points,
             self._air_attenuation, self.n_bins)
-        energy_0_dir = ee_order._add_directional(
+        energy_0_dir = _add_directional(
             energy_0, source_position,
             patches_center, self._n_bins, patch_to_wall_ids,
             sources, receivers,
@@ -216,12 +215,12 @@ class DirectionalRadiosityFast():
         if self._energy_exchange_etc is None or recalculate:
             if max_reflection_order < 1:
                 self._energy_exchange_etc = \
-                    ee_order.energy_exchange_init_energy(
+                    _energy_exchange_init_energy(
                         n_samples, energy_0_dir, distance_0,
                         speed_of_sound, etc_time_resolution,
                         )
             else:
-                self._energy_exchange_etc = ee_order.energy_exchange(
+                self._energy_exchange_etc = _energy_exchange(
                     n_samples, energy_0_dir, distance_0, distance_i_j,
                     self._form_factors_tilde,
                     speed_of_sound, etc_time_resolution,
@@ -310,7 +309,7 @@ class DirectionalRadiosityFast():
             receivers_array = np.array(
                 [s.cartesian for s in self._brdf_outgoing_directions])
 
-            receiver_idx = geometry.get_scattering_data_receiver_index(
+            receiver_idx = form_factor.get_scattering_data_receiver_index(
                 patches_center, receiver_pos[i], receivers_array,
                 self._patch_to_wall_ids)
 
@@ -324,7 +323,7 @@ class DirectionalRadiosityFast():
             if propagation_fx:
                 # accumulate the patch energies towards the receiver
                 # with atmospheric effects (delay, atmospheric attenuation)
-                histogram_out[i] = ee_order._collect_receiver_energy(
+                histogram_out[i] = _collect_receiver_energy(
                         E_matrix,
                         np.linalg.norm(patches_receiver_distance, axis=1),
                                     self.speed_of_sound,
@@ -412,123 +411,6 @@ class DirectionalRadiosityFast():
             assert (self._frequencies == frequencies).all(), \
                 "Frequencies do not match"
 
-def get_scattering_data_receiver_index(
-        pos_i:np.ndarray, pos_j:np.ndarray,
-        receivers:np.ndarray, wall_id_i:np.ndarray,
-        ):
-    """Get scattering data depending on previous, current and next position.
-
-    Parameters
-    ----------
-    pos_i : np.ndarray
-        current position of shape (3)
-    pos_j : np.ndarray
-        next position of shape (3)
-    receivers : np.ndarray
-        receiver directions of all walls of shape (n_walls, n_receivers, 3)
-    wall_id_i : np.ndarray
-        current wall id to get write directional data
-
-    Returns
-    -------
-    scattering_factor: float
-        scattering factor from directivity
-
-    """
-    n_patches = pos_i.shape[0] if pos_i.ndim > 1 else 1
-    receiver_idx = np.empty((n_patches), dtype=np.int64)
-
-    for i in range(n_patches):
-        difference_receiver = pos_i[i]-pos_j
-        difference_receiver /= np.linalg.norm(
-            difference_receiver)
-        receiver_idx[i] = np.argmin(np.sum(
-            (receivers[wall_id_i[i], :]-difference_receiver)**2, axis=-1),
-            axis=-1)
-
-
-    return receiver_idx
-
-
-def get_scattering_data(
-        pos_h:np.ndarray, pos_i:np.ndarray, pos_j:np.ndarray,
-        sources:np.ndarray, receivers:np.ndarray, wall_id_i:np.ndarray,
-        scattering:np.ndarray, scattering_index:np.ndarray):
-    """Get scattering data depending on previous, current and next position.
-
-    Parameters
-    ----------
-    pos_h : np.ndarray
-        previous position of shape (3)
-    pos_i : np.ndarray
-        current position of shape (3)
-    pos_j : np.ndarray
-        next position of shape (3)
-    sources : np.ndarray
-        source directions of all walls of shape (n_walls, n_sources, 3)
-    receivers : np.ndarray
-        receiver directions of all walls of shape (n_walls, n_receivers, 3)
-    wall_id_i : np.ndarray
-        current wall id to get write directional data
-    scattering : np.ndarray
-        scattering data of shape (n_scattering, n_sources, n_receivers, n_bins)
-    scattering_index : np.ndarray
-        index of the scattering data of shape (n_walls)
-
-    Returns
-    -------
-    scattering_factor: float
-        scattering factor from directivity
-
-    """
-    difference_source = pos_h-pos_i
-    difference_receiver = pos_i-pos_j
-
-    difference_source /= np.linalg.norm(difference_source)
-    difference_receiver /= np.linalg.norm(difference_receiver)
-    source_idx = np.argmin(np.sum(
-        (sources[wall_id_i, :, :]-difference_source)**2, axis=-1))
-    receiver_idx = np.argmin(np.sum(
-        (receivers[wall_id_i, :]-difference_receiver)**2, axis=-1))
-    return scattering[scattering_index[wall_id_i],
-        source_idx, receiver_idx, :]
-
-
-def get_scattering_data_source(
-        pos_h:np.ndarray, pos_i:np.ndarray,
-        sources:np.ndarray, wall_id_i:np.ndarray,
-        scattering:np.ndarray, scattering_index:np.ndarray):
-    """Get scattering data depending on previous, current position.
-
-    Parameters
-    ----------
-    pos_h : np.ndarray
-        previous position of shape (3)
-    pos_i : np.ndarray
-        current position of shape (3)
-    sources : np.ndarray
-        source directions of all walls of shape (n_walls, n_sources, 3)
-    wall_id_i : np.ndarray
-        current wall id to get write directional data
-    scattering : np.ndarray
-        scattering data of shape (n_scattering, n_sources, n_receivers, n_bins)
-    scattering_index : np.ndarray
-        index of the scattering data of shape (n_walls)
-
-    Returns
-    -------
-    scattering_factor: float
-        scattering factor from directivity
-
-    """
-    difference_source = pos_h-pos_i
-    difference_source /= np.linalg.norm(difference_source)
-    source_idx = np.argmin(np.sum(
-        (sources[wall_id_i, :, :]-difference_source)**2, axis=-1))
-    return scattering[scattering_index[wall_id_i], source_idx]
-
-
-
 
     @property
     def n_bins(self):
@@ -611,7 +493,6 @@ def get_scattering_data_source(
         return self._speed_of_sound
 
 
-
 def _rotate_coords_to_normal(
         wall_normal:np.ndarray, wall_up_vector:np.ndarray,
         sources:pf.Coordinates, receivers:pf.Coordinates):
@@ -629,10 +510,204 @@ def _rotate_coords_to_normal(
     sources_cp.radius = 1
     return sources_cp, receivers_cp
 
+def _add_directional(
+        energy_0, source_position: np.ndarray,
+        patches_center: np.ndarray, n_bins:float, patch_to_wall_ids:np.ndarray,
+        sources: np.ndarray, receivers: np.ndarray,
+        scattering: np.ndarray, scattering_index: np.ndarray):
+    """Add scattering and absorption to the initial energy from the source.
+
+    Parameters
+    ----------
+    energy_0 : np.ndarray
+        energy of all patches of shape (n_patches, n_bins)
+    source_position : np.ndarray
+        source position of shape (3,)
+    patches_center : np.ndarray
+        center of all patches of shape (n_patches, 3)
+    n_bins : float
+        number of frequency bins.
+    patch_to_wall_ids : np.ndarray
+        indexes from each patch to the wall of shape (n_patches)
+    sources : np.ndarray
+        source positions of shape (n_walls, n_sources, 3)
+    receivers : np.ndarray
+        receiver positions of shape (n_walls, n_receivers, 3)
+    scattering : np.ndarray
+        scattering data of shape (n_walls, n_sources, n_receivers, n_bins)
+    scattering_index : np.ndarray
+        mapping from the wall id to scattering database index (n_walls)
+
+    Returns
+    -------
+    energy : np.ndarray
+        energy of all patches of shape (n_patches, n_directions, n_bins)
+
+    """
+    n_patches = patches_center.shape[0]
+    n_directions = receivers.shape[1]
+    energy_0_directivity = np.zeros((n_patches, n_directions, n_bins))
+    for i in prange(n_patches):
+        wall_id_i = int(patch_to_wall_ids[i])
+        scattering_factor = form_factor.get_scattering_data_source(
+            source_position, patches_center[i],
+            sources, wall_id_i, scattering, scattering_index)
+
+        energy_0_directivity[i, :, :] = energy_0[i] \
+            * scattering_factor
+
+    return energy_0_directivity
+
+
+def _energy_exchange_init_energy(
+        n_samples, energy_0_directivity, distance_0,
+        speed_of_sound, histogram_time_resolution):
+    """Calculate energy exchange between patches.
+
+    Parameters
+    ----------
+    n_samples : int
+        number of samples of the histogram.
+    energy_0_directivity : np.ndarray
+        init energy of all patches of shape (n_patches, n_directions, n_bins)
+    distance_0 : np.ndarray
+        distance from the source to all patches of shape (n_patches)
+    speed_of_sound : float
+        speed of sound in m/s.
+    histogram_time_resolution : float
+        time resolution of the histogram in s.
+
+    Returns
+    -------
+    E_matrix_total : np.ndarray
+        energy of all patches of shape
+        (n_patches, n_directions, n_bins, n_samples)
+
+    """
+    n_patches = energy_0_directivity.shape[0]
+    n_directions = energy_0_directivity.shape[1]
+    n_bins = energy_0_directivity.shape[2]
+    E_matrix_total = np.zeros((n_patches, n_directions, n_bins, n_samples))
+    for i in prange(n_patches):
+        n_delay_samples = int(
+            distance_0[i]/speed_of_sound/histogram_time_resolution)
+        E_matrix_total[i, :, :, n_delay_samples] += energy_0_directivity[i]
+    return E_matrix_total
+
+
+def _energy_exchange(
+        n_samples, energy_0_directivity, distance_0, distance_ij,
+        form_factors_tilde,
+        speed_of_sound, histogram_time_resolution, max_order, visible_patches):
+    """Calculate energy exchange between patches.
+
+    Parameters
+    ----------
+    n_samples : int
+        number of samples of the histogram.
+    energy_0_directivity : np.ndarray
+        init energy of all patches of shape (n_patches, n_directions, n_bins)
+    distance_0 : np.ndarray
+        distance from the source to all patches of shape (n_patches)
+    distance_ij : np.ndarray
+        distance between all patches of shape (n_patches, n_patches)
+    form_factors_tilde : np.ndarray
+        form factors of shape (n_patches, n_patches, n_directions, n_bins)
+    speed_of_sound : float
+        speed of sound in m/s.
+    histogram_time_resolution : float
+        time resolution of the histogram in s.
+    max_order : int
+        maximum order of reflections.
+    visible_patches : np.ndarray
+        indexes of all visible patches of shape (n_visible, 2)
+
+    Returns
+    -------
+    E_matrix_total : np.ndarray
+        energy of all patches of shape
+        (n_patches, n_directions, n_bins, n_samples)
+
+    """
+    n_patches = form_factors_tilde.shape[0]
+    n_directions = form_factors_tilde.shape[2]
+    n_bins = energy_0_directivity.shape[2]
+    form_factors_tilde = form_factors_tilde[..., np.newaxis]
+    E_matrix_total  = _energy_exchange_init_energy(
+        n_samples, energy_0_directivity, distance_0, speed_of_sound,
+        histogram_time_resolution)
+    E_matrix = np.zeros((2, n_patches, n_directions, n_bins, n_samples))
+    E_matrix[0] += E_matrix_total
+    if max_order == 0:
+        return E_matrix_total
+    for k in range(max_order):
+        current_index = (1+k) % 2
+        E_matrix[current_index, :, :, :] = 0
+        for ii in range(visible_patches.shape[0]):
+            for jj in range(2):
+                if jj == 0:
+                    i = visible_patches[ii, 0]
+                    j = visible_patches[ii, 1]
+                else:
+                    j = visible_patches[ii, 0]
+                    i = visible_patches[ii, 1]
+                n_delay_samples = int(
+                    distance_ij[i, j]/speed_of_sound/histogram_time_resolution)
+                if n_delay_samples > 0:
+                    E_matrix[current_index, j, :, :, n_delay_samples:] += \
+                        form_factors_tilde[i, j] * E_matrix[
+                            current_index-1, i, :, :, :-n_delay_samples]
+                else:
+                    E_matrix[current_index, j, :, :, :] += form_factors_tilde[
+                        i, j] * E_matrix[current_index-1, i, :, :, :]
+        E_matrix_total += E_matrix[current_index]
+    return E_matrix_total
+
+
+def _collect_receiver_energy(
+        E_matrix_total, patch_receiver_distance,
+        speed_of_sound, histogram_time_resolution, air_attenuation):
+    """Collect the energy at the receiver.
+
+    Parameters
+    ----------
+    E_matrix_total : np.ndarray
+        energy of all patches of shape
+        (n_patches, n_directions, n_bins, n_samples)
+    patch_receiver_distance : np.ndarray
+        distance from the patch to the receiver of shape (n_patches)
+    speed_of_sound : float
+        speed of sound in m/s.
+    histogram_time_resolution : float
+        time resolution of the histogram in s.
+    air_attenuation : np.ndarray
+        air attenuation factor for each frequency bin of shape (n_bins)
+
+    Returns
+    -------
+    ir : np.ndarray
+        impulse response of shape (n_samples, n_bins)
+
+    """
+    E_mat_out = np.zeros_like(E_matrix_total)
+    n_patches = E_matrix_total.shape[0]
+    n_bins = E_matrix_total.shape[1]
+
+    for i in prange(n_patches):
+        n_delay_samples = int(np.ceil(
+            patch_receiver_distance[i]/speed_of_sound/histogram_time_resolution))
+        for j in range(n_bins):
+            E_mat_out[i,j] = np.roll(
+                E_matrix_total[i,j]*np.exp(-air_attenuation[j]*patch_receiver_distance[i]),
+                n_delay_samples)
+
+    return E_mat_out
 
 if numba is not None:
-    get_scattering_data_receiver_index = numba.njit()(
-        get_scattering_data_receiver_index)
-    get_scattering_data = numba.njit()(get_scattering_data)
-    get_scattering_data_source = numba.njit()(get_scattering_data_source)
- 
+    _add_directional = numba.njit(parallel=True)(_add_directional)
+    _energy_exchange_init_energy = numba.njit()(_energy_exchange_init_energy)
+    _collect_receiver_energy = numba.njit()(_collect_receiver_energy)
+    _energy_exchange = numba.njit()(_energy_exchange)
+
+
+
